fixed unit tests

+++ app/schemas/added_by_rel_schema.py
@@ -8,5 +8,5 @@ def check_epoch_is_valid(data):
 
 
 class AddedByRelSchema(FRPMSchema):
-    epoch = fields.Int(required=True, allow_none=False,
+    epoch = fields.Float(required=True, allow_none=False,
                        description="Unix timestamp: time recorded as number of miliseconds past 01.01.1970")
